+++ src/index.ts
@@ -1,14 +1,14 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 import * as program from "commander";
 import { generateAPIClient } from "./lib";
-import { convertEnums } from "./lib/convertEnums";
+import { convertEnums } from "./lib/commands/convertEnums";
 
 const version = require("../package.json").version;
 program.version(version);
 program.option("-v, --version", "Output current version");
 
 program
-    .command("gen-agent <source> [destination]")
+    .command("gen-agent <sourceFile> <destinationFile>")
     .description("Generate API Agent")
     .action((source, destination) => {
         generateAPIClient(source, destination);
@@ -16,7 +16,7 @@ program
 
 program
     .command("convert-enums <sourceDir> <destinationDir>")
-    .description("clone a repository into a newly created directory")
+    .description("Generate OpenAPI enums from Typescript enums")
     .action((sourceDir, destinationDir) => {
         convertEnums(sourceDir, destinationDir);
     });